labvision/requirements.txt
paramiko==2.7.1
matplotlib==3.1.1
wandb==0.8.35
scipy==1.3.1
torch==1.3.0
opencv_python==4.1.0.25
numpy==1.16.5
torchvision==0.4.1
imageio==2.8.0
Pillow==7.1.2
